Move web3 mocking for vault tests into its own test file
